moved unassigned students to its own component. Started basic structure to furniture modal component
