+++ client/src/components/SeatingChart/AddFurnitureToClassroom.jsx
@@ -1,0 +1,13 @@
+import { useState, useRef } from "react";
+import { motion, useMotionValue } from "framer-motion";
+
+const AddFurnitureToClassroom = ({ }) => {
+  
+  return (
+    <>
+      
+    </>
+  );
+};
+
+export default AddFurnitureToClassroom;
