+++ client/src/components/SeatingChart/AddStudentsToClass.jsx
@@ -1,7 +1,8 @@
 import { useEffect, useState } from "react";
 import SampleAvatar from '../../images/Sample_Avatar.png'
+import UnassignedStudent from "./UnassignedStudent";
 
-const AddStudentModal = ({unassignedStudents, students, onClose,onConfirm}) => {
+const AddStudentModal = ({unassignedStudents, students, onClose, onConfirm}) => {
 
 // FIXME: might still need, could switch the x and y coords to null 
     // setStudentPositions((prevPositions) => ({
@@ -12,6 +13,8 @@ const AddStudentModal = ({unassignedStudents, students, onClose,onConfirm}) => {
     //       assigned: false,
     //     },
     //   }));
+
+    // TODO: Need to be able to add students back to the assigned students array
 
     const [selectedStudent, setSelectedStudent] = useState("")
 
@@ -51,32 +54,7 @@ const AddStudentModal = ({unassignedStudents, students, onClose,onConfirm}) => {
     <>
         <div className="absolute top-60 left-[8.5%] z-10 h-[648px] w-[686px] bg-notebookPaper border-sandwich border-4 p-10 rounded">
             <div className="flex flex-row flex-wrap">
-            {unassignedStudents.map((studentObj, index) => {
-                const unassignedStudent = students.find(
-                    (student) => student._id === studentObj.student
-                );
-
-                if (unassignedStudent) {
-                    return (
-                    <div
-                        id={`motion-div-${unassignedStudent._id}`}
-                        key={`unassigned-${index}`}
-                        className={`m-2 bg-${unassignedStudent.borderColorClass} h-[99px] w-[99px] ${
-                            selectedStudent.includes(studentObj.student) 
-                            ? "opacity-50"
-                            : ""
-                        } rounded-2xl`}
-                        onClick={() => addSelectedStudent(studentObj)}
-                    >
-                        <div className="flex w-full justify-center h-full items-center">
-                            <img className="flex object-cover h-[84px] w-[84px] rounded-2xl" src={SampleAvatar} />
-                        </div>
-                    </div>
-                    );
-                } else {
-                    return null;
-                }
-                })}
+                <UnassignedStudent unassignedStudents={unassignedStudents} students={students} selectedStudent={selectedStudent} addSelectedStudent={addSelectedStudent}/>
             </div>
             <div className="flex h-[400px] w-full justify-center items-end">
                 <button className="w-32 h-10 bg-lightLavender self-end" onClick={() => {
